requirements.txt
appdirs==1.4.3
Babel==2.4.0
bcrypt==3.1.3
beautifulsoup4==4.6.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.3
Django==1.10
django-filter==1.0.4
django-phonenumber-field==1.3.0
django-twilio==0.8.0
djangorestframework==3.6.3
httpie==0.9.9
idna==2.5
Markdown==2.6.8
olefile==0.44
packaging==16.8
phonenumberslite==8.5.0
Pillow==4.2.1
pycparser==2.17
Pygments==2.2.0
PyJWT==1.5.0
pyparsing==2.2.0
PySocks==1.6.7
pytesseract==0.1.7
pytz==2017.2
requests==2.17.3
six==1.10.0
twilio==6.3.0
urllib3==1.21.1
